+++ metadata.rb
@@ -10,6 +10,8 @@ long_description 'Installs/Configures Divvy'
 version          '0.1.1'
 
 depends          'mac-app-store', '~> 0.1'
+depends          'windows', '~> 1.36'
 
 supports         'mac_os_x'
+supports         'windows'
 # rubocop:enable SingleSpaceBeforeFirstArg