+++ recipes/default.rb
@@ -18,8 +18,8 @@
 # limitations under the License.
 #
 
-include_recipe 'mac-app-store'
+include_recipe 'mac-app-store' if node['platform'] == 'mac_os_x'
 
 divvy_app 'default' do
-  action :install
+  action [:install, :run]
 end