+++ test/integration/mac_direct_download/serverspec/localhost/app_spec.rb
@@ -1,0 +1,17 @@
+# Encoding: UTF-8
+
+require_relative '../spec_helper'
+
+describe 'Divvy app' do
+  describe file('/Applications/Divvy.app') do
+    it 'is present on the filesystem' do
+      expect(subject).to be_directory
+    end
+  end
+
+  describe process('Divvy') do
+    it 'is running' do
+      expect(subject).to be_running
+    end
+  end
+end