+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Divvy Cookbook CHANGELOG
 
 v?.?.? (????-??-??)
 -------------------
+- Add support for non-App Store direct downloads for OS X
+- Add support for Windows
+- Grant required Accessibility privileges automatically on OS X
+- Add a `:run` action and start Divvy after it's installed
 
 v0.1.0 (2015-04-25)
 -------------------